refactor: Removed TestCore from Libraries (#858)
Energinet.Charges.TestCore : moved parts to own dependencies.
