+++ .github/workflows/clients-ci.yml
@@ -32,6 +32,7 @@ jobs:
           github_token: ${{ github.token }}
           paths: '[
             "source/Energinet.Charges.Libraries/**",
+            "source/GreenEnergyHub.Charges/source/GreenEnergyHub.Charges.TestCore/**",
             ".github/workflows/clients-ci.yml",
             ".github/actions/dotnet-build-and-test/**",
             ".github/actions/azure-core-tools-install/**"
