+++ source/Energinet.Charges.Libraries/Energinet.DataHub.Charges.Libraries.sln
@@ -34,11 +34,11 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Energinet.DataHub.Charges.C
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Energinet.DataHub.Charges.Clients.IntegrationTests", "source\Energinet.DataHub.Charges.Clients.IntegrationTests\Energinet.DataHub.Charges.Clients.IntegrationTests.csproj", "{5BFF36A2-AD09-4213-9A7F-B1CAD8298822}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Energinet.DataHub.Charges.TestCore", "source\Energinet.DataHub.Charges.TestCore\Energinet.DataHub.Charges.TestCore.csproj", "{02FF3F12-886A-45D5-9A4D-E1EA24EE899C}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Energinet.DataHub.Charges.Clients.SimpleInjector", "source\Energinet.DataHub.Charges.Clients.SimpleInjector\Energinet.DataHub.Charges.Clients.SimpleInjector.csproj", "{9BE290B0-8A53-4157-A822-BAA05B0627F4}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Energinet.DataHub.Charges.Clients.SimpleInjectorTests", "source\Energinet.DataHub.Charges.Clients.SimpleInjectorTests\Energinet.DataHub.Charges.Clients.SimpleInjectorTests.csproj", "{12C264C7-A6AD-4644-9681-1FEB9DE4F230}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GreenEnergyHub.Charges.TestCore", "..\GreenEnergyHub.Charges\source\GreenEnergyHub.Charges.TestCore\GreenEnergyHub.Charges.TestCore.csproj", "{4A55C6A6-1496-4470-B2B8-EA179A6DA7BD}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -98,18 +98,6 @@ Global
 		{5BFF36A2-AD09-4213-9A7F-B1CAD8298822}.Release|x64.Build.0 = Release|Any CPU
 		{5BFF36A2-AD09-4213-9A7F-B1CAD8298822}.Release|x86.ActiveCfg = Release|Any CPU
 		{5BFF36A2-AD09-4213-9A7F-B1CAD8298822}.Release|x86.Build.0 = Release|Any CPU
-		{02FF3F12-886A-45D5-9A4D-E1EA24EE899C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{02FF3F12-886A-45D5-9A4D-E1EA24EE899C}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{02FF3F12-886A-45D5-9A4D-E1EA24EE899C}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{02FF3F12-886A-45D5-9A4D-E1EA24EE899C}.Debug|x64.Build.0 = Debug|Any CPU
-		{02FF3F12-886A-45D5-9A4D-E1EA24EE899C}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{02FF3F12-886A-45D5-9A4D-E1EA24EE899C}.Debug|x86.Build.0 = Debug|Any CPU
-		{02FF3F12-886A-45D5-9A4D-E1EA24EE899C}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{02FF3F12-886A-45D5-9A4D-E1EA24EE899C}.Release|Any CPU.Build.0 = Release|Any CPU
-		{02FF3F12-886A-45D5-9A4D-E1EA24EE899C}.Release|x64.ActiveCfg = Release|Any CPU
-		{02FF3F12-886A-45D5-9A4D-E1EA24EE899C}.Release|x64.Build.0 = Release|Any CPU
-		{02FF3F12-886A-45D5-9A4D-E1EA24EE899C}.Release|x86.ActiveCfg = Release|Any CPU
-		{02FF3F12-886A-45D5-9A4D-E1EA24EE899C}.Release|x86.Build.0 = Release|Any CPU
 		{9BE290B0-8A53-4157-A822-BAA05B0627F4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{9BE290B0-8A53-4157-A822-BAA05B0627F4}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{9BE290B0-8A53-4157-A822-BAA05B0627F4}.Debug|x64.ActiveCfg = Debug|Any CPU
@@ -134,6 +122,18 @@ Global
 		{12C264C7-A6AD-4644-9681-1FEB9DE4F230}.Release|x64.Build.0 = Release|Any CPU
 		{12C264C7-A6AD-4644-9681-1FEB9DE4F230}.Release|x86.ActiveCfg = Release|Any CPU
 		{12C264C7-A6AD-4644-9681-1FEB9DE4F230}.Release|x86.Build.0 = Release|Any CPU
+		{4A55C6A6-1496-4470-B2B8-EA179A6DA7BD}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{4A55C6A6-1496-4470-B2B8-EA179A6DA7BD}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{4A55C6A6-1496-4470-B2B8-EA179A6DA7BD}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{4A55C6A6-1496-4470-B2B8-EA179A6DA7BD}.Debug|x64.Build.0 = Debug|Any CPU
+		{4A55C6A6-1496-4470-B2B8-EA179A6DA7BD}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{4A55C6A6-1496-4470-B2B8-EA179A6DA7BD}.Debug|x86.Build.0 = Debug|Any CPU
+		{4A55C6A6-1496-4470-B2B8-EA179A6DA7BD}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{4A55C6A6-1496-4470-B2B8-EA179A6DA7BD}.Release|Any CPU.Build.0 = Release|Any CPU
+		{4A55C6A6-1496-4470-B2B8-EA179A6DA7BD}.Release|x64.ActiveCfg = Release|Any CPU
+		{4A55C6A6-1496-4470-B2B8-EA179A6DA7BD}.Release|x64.Build.0 = Release|Any CPU
+		{4A55C6A6-1496-4470-B2B8-EA179A6DA7BD}.Release|x86.ActiveCfg = Release|Any CPU
+		{4A55C6A6-1496-4470-B2B8-EA179A6DA7BD}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -145,9 +145,9 @@ Global
 		{68B83582-7A4C-49F3-AA04-36D390BDA5F0} = {D79C7D37-ADE1-42D3-9FED-1E5D784E6E7D}
 		{B2D07074-0796-4553-AF4D-768AAC624528} = {68B83582-7A4C-49F3-AA04-36D390BDA5F0}
 		{5BFF36A2-AD09-4213-9A7F-B1CAD8298822} = {68B83582-7A4C-49F3-AA04-36D390BDA5F0}
-		{02FF3F12-886A-45D5-9A4D-E1EA24EE899C} = {68B83582-7A4C-49F3-AA04-36D390BDA5F0}
 		{9BE290B0-8A53-4157-A822-BAA05B0627F4} = {2BF2F6B8-E02B-4D7F-99BC-D28B3B3643A9}
 		{12C264C7-A6AD-4644-9681-1FEB9DE4F230} = {68B83582-7A4C-49F3-AA04-36D390BDA5F0}
+		{4A55C6A6-1496-4470-B2B8-EA179A6DA7BD} = {68B83582-7A4C-49F3-AA04-36D390BDA5F0}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {AFC1A6A3-7235-4DB1-A2E4-35B72C1FEA28}
