--- source/Energinet.Charges.Libraries/source/Energinet.DataHub.Charges.TestCore/Energinet.DataHub.Charges.TestCore.csproj
@@ -1,44 +1,0 @@
-﻿<!--
-Copyright 2020 Energinet DataHub A/S
-
-Licensed under the Apache License, Version 2.0 (the "License2");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
--->
-<Project Sdk="Microsoft.NET.Sdk">
-
-    <PropertyGroup>
-        <TargetFramework>net5.0</TargetFramework>
-        <Nullable>enable</Nullable>
-        <IsPackable>false</IsPackable>
-        <TreatWarningsAsErrors>true</TreatWarningsAsErrors>
-        <RunAnalyzersDuringBuild>true</RunAnalyzersDuringBuild>
-        <LangVersion>9</LangVersion>
-        <OutputType>Library</OutputType>
-        <IsTestProject>false</IsTestProject>
-    </PropertyGroup>
-
-    <ItemGroup>
-      <PackageReference Include="AutoFixture" Version="4.17.0" />
-      <PackageReference Include="AutoFixture.AutoMoq" Version="4.17.0" />
-      <PackageReference Include="AutoFixture.Xunit2" Version="4.17.0" />
-      <PackageReference Include="NodaTime" Version="3.0.9" />
-      <PackageReference Include="MicroElements.AutoFixture.NodaTime" Version="1.0.0" />
-      <PackageReference Include="System.ComponentModel.TypeConverter" Version="4.3.0" />
-      <PackageReference Include="xunit.extensibility.core" Version="2.4.1" />
-      <PackageReference Include="xunit" Version="2.4.1" />
-      <PackageReference Update="StyleCop.Analyzers" Version="1.2.0-beta.354">
-        <PrivateAssets>all</PrivateAssets>
-        <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
-      </PackageReference>
-    </ItemGroup>
-
-</Project>
