+++ source/GreenEnergyHub.Charges/source/GreenEnergyHub.Charges.IntegrationTests/Fixtures/ChargesFunctionAppFixture.cs
@@ -26,8 +26,8 @@ using Energinet.DataHub.Core.FunctionApp.TestCommon.ServiceBus.ResourceProvider;
 using Energinet.DataHub.Core.TestCommon.Diagnostics;
 using Energinet.DataHub.MessageHub.IntegrationTesting;
 using GreenEnergyHub.Charges.FunctionHost.Common;
+using GreenEnergyHub.Charges.IntegrationTests.Database;
 using GreenEnergyHub.Charges.IntegrationTests.TestCommon;
-using GreenEnergyHub.Charges.TestCore.Database;
 using Microsoft.Extensions.Configuration;
 
 namespace GreenEnergyHub.Charges.IntegrationTests.Fixtures
