+++ source/GreenEnergyHub.Charges/source/GreenEnergyHub.Charges.Tests/Infrastructure/Integration/Mappers/ChargePricesUpdatedOutboundMapperTests.cs
@@ -17,7 +17,7 @@ using System.Diagnostics.CodeAnalysis;
 using GreenEnergyHub.Charges.Application.Charges.Acknowledgement;
 using GreenEnergyHub.Charges.Infrastructure.Integration.Mappers;
 using GreenEnergyHub.Charges.TestCore.Attributes;
-using GreenEnergyHub.Charges.TestCore.Protobuf;
+using GreenEnergyHub.Charges.Tests.Protobuf;
 using Xunit;
 using Xunit.Categories;
